Everything related to search and querying is moved in to
dbnp.studycatpuring.query (for controllers) and
views.query

The reason for moving this is that apparently searchable has been found to overwrite files in the searchable directory.
searchable/ is still kept but not used any longer.

Furthermore, for the time being, the advanced search option is removed because it is not part of the prototype.