- resolved issue #127 ( https://trac.nbic.nl/gscf/ticket/127 ): new terms do not show in termSelect element after adding them
This bug is caused by the Ontology being updated on the ncbo side. We already have one version in the database (1.71) which has only one term linked (glucose). When you tried to add other terms, these terms came from a newer version of the Ontology (1.72) which did not yet exist in the database so a new Ontology was created. The Terms benzo[a]pyrene, benzene and benzenediols were added to this Ontology. 

Resolution: before adding a new ontology, a lookup is performed on name. If this returns an existing ontology, the existing ontology is updated with the updated information of the new version of the ontology. This sollution is not ideal, but good enough for now...