/views/error.gsp: made this view less error-prone
Due to how we are trying to bypass the AjaxFlow plugin in order to be
able to deliver attachments to the client without breaking the flow, we
trigger an illegal state exception. Grails tries to render
/views/error.gsp, however, an error occurs on the line in question.
After that, we write our attachment to the response and send it back to
the client. This works fine, as long as the line in question doesn't
block execution. For that reason, some questionmarks have been added to
this line.