implemented new API calls
- getFieldsForEntity
- getFieldsForEntityWithTemplate
- createEntityWithTemplate

The latter is a bit harder, as it also has to make sure any cascaded
relationships are properly set. In order to do so, the caller should
provide any relationships by setting an entity token. For example, by
passing a studyToken it is implied that the entity should be added to a
study. If a (has many) relationship exists this will be handled properly

Note that this should be tested!

Another difficulty is access rights. If a has many relationship is
given, the api currently would check if you have access to the
instance, not the instance's relationships. So if you add an assay, it
should check for belongsTo relationships for additional authorisation
checks (e.g. belongsTo Study --> check study for authorisation as well)