Implemented search rest method in SAM.
In order to reuse most of the functionality already present in GSCF,
the query classes were moved to the dbxpBase plugin. Within SAM
an extension of those classes is made and those classes are used
to search. In the end, the SAM classes are much simpler, but still
benefit from the structure of the query classes.