Refactored study associations into sets to make addition of new entities easier
Instead of storing lists of e.g. subjects with an index in that list, we will now
use sets, and sort the entities by when retrieving them from the database.

This will make importing the entities much faster, as grails doesn't need to
retrieve all entities for a study from the database when adding a single new entity.